Update SetIdAsLong to use Int64 for id retrieval
Modified the `SetIdAsLong` method in the `Helpers` class to retrieve the `id` from the `BsonDocument` as an `Int64` instead of `Int32`. This change ensures that the `id` is explicitly handled as a long integer, accommodating larger values as needed.

+++ src/RepoRanger/Helpers.cs
@@ -27,7 +27,7 @@ public static class Helpers
         // Ensure `id` is stored as a long (Int64)
         if (item.TryGetElement("id", out BsonElement idElement) && idElement.Value.IsInt32)
         {
-            long repoId = idElement.Value.AsInt32; // Explicitly retrieve as long
+            long repoId = idElement.Value.AsInt64; // Explicitly retrieve as long
             item["id"] = new BsonInt64(repoId); // Ensure it's stored as BsonInt64
         }
     }
